mmu: add mmu_test_20
Test if tlbie (plus ptesync) takes effect if used in the same
Translation Block of a subsequent read.

Some other changes were made to be able to unmap an address
without calling tlbie_va(), so that it can be used close enough
to the instruction that tries to read from the invalidated address.