mmu: Introduce PA/VA macros
These macros add a base address value to its argument, to make
sure the resulting address will always be above the base address.
This is needed, especially for the physical addresses, to avoid
overwriting code or data areas.

This is in preparation to add support for 4-level POWER9 Radix
Tree configuration, that uses larger tables.

The physical/virtual addresses used in the tests were also
adjusted to make them compatible with 64K pages.

Signed-off-by: Leandro Lupori <leandro.lupori@eldorado.org.br>